requirements.txt
aiogram
openai==1.93.0
chromadb
python-dotenv
numpy
Django
django-cors-headers
django-jazzmin
djangorestframework
djangorestframework-simplejwt
drf-spectacular
jsonschema
psycopg2-binary
requests
django-admin-list-filter-dropdown
django-admin-rangefilter
django-admin-rangefilter
django-filter
geopy
langchain
langchain-community
langchain-openai